requirements.txt
requests
beautifulsoup4
streamlit
torch
numpy
transformers
google-generativeai
scikit-learn
python-docx
PyPDF2
